Shift around email logic, update specs, move santizer to lib/ add 3 features, add templates to yml,

+++ modules/burials/app/controllers/burials/v0/claims_controller.rb
@@ -3,6 +3,7 @@
 require 'burials/benefits_intake/submit_claim_job'
 require 'burials/monitor'
 require 'common/exceptions/validation_errors'
+require 'persistent_attachments/sanitizer'
 
 module Burials
   module V0
@@ -95,7 +96,10 @@ module Burials
         claim.process_attachments!
       rescue => e
         monitor.track_process_attachment_error(in_progress_form, claim, current_user)
-        claim.destroy! # Handle deletion of the claim if attachments processing fails
+        if Flipper.enabled?(:burial_persistent_attachment_error_email_notification)
+          PersistentAttachments::Sanitizer.new.sanitize_attachments(claim, in_progress_form)
+          claim.destroy! # Handle deletion of the claim if attachments processing fails
+        end
         raise e
       end
